feat(compiler): Calculate struct size and alignment internally (#64)

+++ src/alumina-boot/src/ast/mod.rs
@@ -704,7 +704,7 @@ pub enum Attribute {
     Cold,
     TestMain,
     Inline,
-    Align(u32),
+    Align(usize),
     Packed,
     Transparent,
     NoInline,
